Split context into a serialization and data source part
